+++ Project-c1120g1/src/main/java/com/c1120g1/adweb/service/AccountService.java
@@ -7,4 +7,6 @@ public interface AccountService {
     Account findByUsername(String username);
 
     String generateCode();
+
+    void sendEmail(String email, String code);
 }
