+++ Project-c1120g1/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/ad_web_db?createDatabaseIfNotExist=true
 spring.datasource.username=root
-spring.datasource.password=tgnaboas5
+spring.datasource.password=hoangnhan
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
